ocr_app/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
pytesseract==0.3.10
Pillow==10.2.0
